backend/requirements.txt
fastapi
uvicorn
black
pylint
pydantic
asyncpg
sqlalchemy
pydantic[email]
psycopg2-binary
python-dotenv
pytz
python-jose==3.3.0
passlib>=1.7.4
python-multipart
bcrypt==4.3.0

